requirements.txt
python-dotenv
lingua-language-detector
# backports.zoneinfo; python_version<'3.9'  # Uncomment if you use zoneinfo and need Python <3.9 support
